// src/app/model/car-property.ts
export interface CarProperty {
    make: string;
    model: string;
    engine: string;
    color: string;
    year: any;
    transmission: string;
    mileage: string;
    fuel: string;
    seller: string;
    images: any[];
    phoneNumber: number;
    description: string;
    price: number;
    time: string;
    uniqueId: number;
    isClicked?: boolean;
    email: string;
}